requirements.txt
aniso8601==9.0.1
attrs==21.4.0
click==7.1.2
Flask==1.1.4
flask-marshmallow==0.14.0
flask-restplus==0.13.0
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-resources==5.4.0
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.3.2
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.26.1
mysqlclient==2.1.0
pyrsistent==0.18.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.29
SQLAlchemy-Utils==0.38.2
Werkzeug==1.0.1
zipp==3.7.0
